Move playback service
