+++ android/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
 
     <!-- Allow background audio playback by registering this service. -->
     <service
-      android:name="com.theoplayer.audio.MediaPlaybackService"
+      android:name="com.theoplayer.media.MediaPlaybackService"
       android:description="@string/background_playback_service_description"
       android:exported="false"
       android:enabled="false"
